requirements.txt
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
colorama==0.4.4
Django==3.2.5
django-extensions==3.1.3
django-pdb==0.6.2
djangorestframework==3.12.4
Faker==8.10.3
flake8==3.9.2
glob2==0.7
iniconfig==1.1.1
Mako==1.1.4
MarkupSafe==2.0.1
mccabe==0.6.1
mixer==7.1.2
packaging==21.0
parse==1.19.0
parse-type==0.5.2
pep8==1.7.1
pluggy==0.13.1
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==2.4.7
pytest==6.2.4
pytest-bdd==4.1.0
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
